refactor: Move expense directory into model package
