+++ src/main/java/br/com/matheusgusmao/incometax/domain/service/DeclarationService.java
@@ -1,6 +1,6 @@
 package br.com.matheusgusmao.incometax.domain.service;
 
-import br.com.matheusgusmao.incometax.domain.expense.DeductibleExpense;
+import br.com.matheusgusmao.incometax.domain.model.expense.DeductibleExpense;
 import br.com.matheusgusmao.incometax.domain.model.declaration.Declaration;
 import br.com.matheusgusmao.incometax.domain.model.income.Income;
 import br.com.matheusgusmao.incometax.infra.exception.custom.EntityAlreadyExistsException;
