+++ src/main/java/br/com/matheusgusmao/incometax/web/dto/expense/ExpenseResponse.java
@@ -1,7 +1,7 @@
 package br.com.matheusgusmao.incometax.web.dto.expense;
 
-import br.com.matheusgusmao.incometax.domain.expense.DeductibleExpense;
-import br.com.matheusgusmao.incometax.domain.expense.ExpenseType;
+import br.com.matheusgusmao.incometax.domain.model.expense.DeductibleExpense;
+import br.com.matheusgusmao.incometax.domain.model.expense.ExpenseType;
 
 import java.math.BigDecimal;
